feat(1263): tweaks
