+++ apps/token/src/routes/staking/home/epoch-data.tsx
@@ -1,25 +1,11 @@
 import { useEffect } from 'react';
 import { AsyncRenderer } from '@vegaprotocol/ui-toolkit';
 import { EpochCountdown } from '../../../components/epoch-countdown';
-import { Schema } from '@vegaprotocol/types';
 import { useNodesQuery } from './__generated___/Nodes';
 import { usePreviousEpochQuery } from './__generated___/PreviousEpoch';
 import { ValidatorTables } from './validator-tables';
-import { NetworkParams, useNetworkParams } from '@vegaprotocol/react-helpers';
 
-export const statusTranslationKey = (status: Schema.ValidatorStatus) => {
-  // Returns a key for translation
-  const statuses = {
-    [Schema.ValidatorStatus.VALIDATOR_NODE_STATUS_ERSATZ]: 'status-ersatz',
-    [Schema.ValidatorStatus.VALIDATOR_NODE_STATUS_PENDING]: 'status-pending',
-    [Schema.ValidatorStatus.VALIDATOR_NODE_STATUS_TENDERMINT]:
-      'status-tendermint',
-  };
-
-  return statuses[status];
-};
-
-export const NodeList = () => {
+export const EpochData = () => {
   // errorPolicy due to vegaprotocol/vega issue 5898
   const { data, error, loading, refetch } = useNodesQuery();
   const { data: previousEpochData, refetch: previousEpochRefetch } =
