+++ apps/token/src/routes/staking/home/validator-tables/index.ts
@@ -1,0 +1,1 @@
+export * from './validator-tables';
