+++ apps/token/src/routes/staking/node/validator-table.tsx
@@ -6,10 +6,22 @@ import { useEnvironment } from '@vegaprotocol/environment';
 import { KeyValueTable, KeyValueTableRow } from '@vegaprotocol/ui-toolkit';
 import { BigNumber } from '../../../lib/bignumber';
 import { formatNumber } from '../../../lib/format-number';
-import { statusTranslationKey } from '../home/node-list';
 import type { StakingNodeFieldsFragment } from './__generated___/Staking';
 import { toBigNum } from '@vegaprotocol/react-helpers';
 import { useAppState } from '../../../contexts/app-state/app-state-context';
+import { Schema } from '@vegaprotocol/types';
+
+const statusTranslationKey = (status: Schema.ValidatorStatus) => {
+  // Returns a key for translation
+  const statuses = {
+    [Schema.ValidatorStatus.VALIDATOR_NODE_STATUS_ERSATZ]: 'status-ersatz',
+    [Schema.ValidatorStatus.VALIDATOR_NODE_STATUS_PENDING]: 'status-pending',
+    [Schema.ValidatorStatus.VALIDATOR_NODE_STATUS_TENDERMINT]:
+      'status-tendermint',
+  };
+
+  return statuses[status];
+};
 
 const ValidatorTableCell = ({
   children,
